Add files via upload
Okay so this is an app built using JS. It s a CRUD app. The user can add a contact, edit a contact and delete a contact as well as refresh the contact list. We used an API created by the Fnb app academy and used the "rootPath" to fetch info from the server.